Split the test lgrIJK - lgrCOORDZCORN
