+++ tests/cpgrid/lgrIJK_test.cpp
@@ -48,11 +48,9 @@
 #include <opm/input/eclipse/Deck/Deck.hpp>
 #include <opm/input/eclipse/EclipseState/EclipseState.hpp>
 #include <opm/input/eclipse/Parser/Parser.hpp>
-#include <opm/io/eclipse/EGrid.hpp> // check if it's needed
 
 
 #include <array>
-#include <filesystem>
 #include <stdexcept>
 #include <string>
 #include <vector>
@@ -244,153 +242,6 @@ SCHEDULE
     BOOST_TEST(lgr1IJK[77][0] == 2);
     BOOST_TEST(lgr1IJK[77][1] == 4);
     BOOST_TEST(lgr1IJK[77][2] == 2);
-
-   
-    const auto [lgrCOORD, lgrZCORN] = Opm::lgrCOORDandZCORN(grid, lgr1_level, lgrCartesianIdxToCellIdx, lgr1IJK);
-    
-    std::cout<< "COORD for LGR with all active cells" << std::endl;
-    for (const auto& coord : lgrCOORD)
-    {
-        std::cout << coord << std::endl;
-    }
-    std::cout<< std::endl;
-}
-
-
-BOOST_AUTO_TEST_CASE(singleCellGrid_easyToTestlgrCOORDandZCORN)
-{
-    // Single-cell-grid (with dimension 1x1x1)
-    // {DX,DY,DZ} = {8,4,2} which makes easy to check the values
-    // of LGR COORD if the LGR dimensions are 8x4x2.
-
-    const std::string deck_string = R"(
-RUNSPEC
-DIMENS
-  1 1 1 /
-GRID
-CARFIN
--- NAME I1-I2 J1-J2 K1-K2 NX NY NZ
-'LGR1'  1  1  1  1  1  1  8  4  2/
-ENDFIN
-DX
-  1*8 /
-DY
-	1*4 /
-DZ
-	1*2 /
-TOPS
-	1*0 /
- ACTNUM
-        1
-        /
-PORO
-  1*0.15 /
-PERMX
-  1*1 /
-COPY
-  PERMX PERMZ /
-  PERMX PERMY /
-/
-EDIT
-OIL
-GAS
-TITLE
-The title
-START
-16 JUN 1988 /
-PROPS
-REGIONS
-SOLUTION
-SCHEDULE
-)";
-
-    Dune::CpGrid grid;
-    // Create the starting grid (before adding LGRs)
-    Opm::Parser parser;
-    const auto deck = parser.parseString(deck_string);
-    Opm::EclipseState ecl_state(deck);
-    Opm::EclipseGrid eclipse_grid = ecl_state.getInputGrid();
-    
-    grid.processEclipseFormat(&eclipse_grid, &ecl_state, false, false, false);
-
-    // Add LGR1 and update grid view
-    const std::vector<std::array<int, 3>> cells_per_dim_vec
-        = {{8, 4, 2}}; // 8x4x2 child cells in x-,y-, and z-direction per ACTIVE parent cell
-    const std::vector<std::array<int, 3>> startIJK_vec
-        = {{0, 0, 0}}; // starts at (0,0,0) in coarse grid - equivalent to (I1-1, J1-1, K1-1) from its CARFIN block
-    const std::vector<std::array<int, 3>> endIJK_vec
-        = {{1, 1, 1}}; // ends at (1,1,1) in coarse grid - equivalent to (I2, J2, K2) from its CARFIN block
-    const std::vector<std::string> lgr_name_vec = {"LGR1"};
-    grid.addLgrsUpdateLeafView(cells_per_dim_vec, startIJK_vec, endIJK_vec, lgr_name_vec);
-
-    // Get LGR level
-    const int lgr1_level = grid.getLgrNameToLevel().at("LGR1");
-    const int numLgrCells = grid.levelGridView(lgr1_level).size(0);
-
-    const auto [lgrCartesianIdxToCellIdx, lgr1IJK] = Opm::lgrIJK(grid, "LGR1");
-    const auto cellIdxToLgrCartesianIdx = grid.currentData()[lgr1_level]->globalCell();
-
-    // Verify the size matches expected elements
-    const int expected_elements = 64; // 1 parent cells into 8x4x2 children each -> 64
-    checkExpectedSize(expected_elements,
-                      numLgrCells,
-                      cellIdxToLgrCartesianIdx.size(),
-                      lgrCartesianIdxToCellIdx.size(),
-                      lgr1IJK.size());
-
-
-    const auto [lgrCOORD, lgrZCORN] = Opm::lgrCOORDandZCORN(grid, lgr1_level, lgrCartesianIdxToCellIdx, lgr1IJK);
-    const int nx = 8;
-    const int ny = 4;
-    const int nz = 2;
-    // Pillars are ordered j*(nx+1) + i, i faster than j, i=0,...,nx, j=0, ..., ny.
-    for (int j = 0;  j < ny+1; ++j) {
-        for (int i = 0; i < nx+1; ++i) {
-            int pillar_idx = (j*6*(nx+1)) + (6*i);
-            BOOST_CHECK_EQUAL( lgrCOORD[pillar_idx +0], i);
-            BOOST_CHECK_EQUAL( lgrCOORD[pillar_idx +1], j);
-            BOOST_CHECK_EQUAL( lgrCOORD[pillar_idx +2], nz);
-            BOOST_CHECK_EQUAL( lgrCOORD[pillar_idx +3], i);
-            BOOST_CHECK_EQUAL( lgrCOORD[pillar_idx +4], j);
-            BOOST_CHECK_EQUAL( lgrCOORD[pillar_idx +5], 0 );
-        }
-    }
-
-    std::cout<< "COORD for LGR with all active cells" << std::endl;
-    for (const auto& coord : lgrCOORD)
-    {
-
-        std::cout << coord << std::endl;
-    }
-    std::cout<< std::endl;
-
-    std::cout<< "ZCORN for LGR with all active cells" << std::endl;
-    for (const auto& zcorn : lgrZCORN)
-    {
-        std::cout << zcorn << std::endl;
-    }
-    std::cout<< std::endl;
-
-    // For a grid with nz layers, ZCORN values are ordered:
-    //
-    //      top layer nz-1
-    //   bottom layer nz-1
-    //      top layer nz-2
-    //   bottom layer nz-2
-    // ...
-    //      top layer 1
-    //   bottom layer 1
-    //      top layer 0
-    //   bottom layer 0
-    for (int k = 0; k < nz; ++k) {
-        for (int j = 0;  j < ny; ++j) {
-            for (int i = 0; i < nx; ++i) {
-                int zcorn_idx =  ((nz-1-k)*8*nx*ny) + (j*4*nx) + (2*i);
-                BOOST_CHECK_EQUAL( lgrZCORN[zcorn_idx], k+1); // top layer zcorn values
-                BOOST_CHECK_EQUAL( lgrZCORN[zcorn_idx + (4*nx*ny)], k); // bottom layer zcorn values
-            }
-        }
-    }
 }
 
 
@@ -459,7 +310,7 @@ SCHEDULE
     //   1  1
     //   1  0
 
-    // Get LGR level
+
     const int lgr1_level = grid.getLgrNameToLevel().at("LGR1");
     // Total active refined cells on the level grid
     const int numLgrCells = grid.levelGridView(lgr1_level).size(0);
@@ -530,16 +381,6 @@ SCHEDULE
     // Accessing inactive (non-existing) cells
     BOOST_CHECK_THROW(lgrCartesianIdxToCellIdx.at(70), std::out_of_range);
     BOOST_CHECK_THROW(lgrCartesianIdxToCellIdx.at(170), std::out_of_range);
-
-    // If a pillar within the LGR block is "inactive," its COORD values are set to
-    // std::numeric_limits<double>::max() to indicate the inactive status
-    const auto [lgrCOORD, lgrZCORN] = Opm::lgrCOORDandZCORN(grid, lgr1_level, lgrCartesianIdxToCellIdx, lgr1IJK);
-    std::cout<< "COORD for LGR with active and inactive cells" << std::endl;
-    for (const auto& coord : lgrCOORD)
-    {
-        std::cout << coord << std::endl;
-    }
-    std::cout << std::endl;
 }
 
 BOOST_AUTO_TEST_CASE(fullInactiveParentCellsBlock_lgrCOORD_throws, *boost::unit_test::disabled())
@@ -600,7 +441,6 @@ SCHEDULE
     //   0  0
     //   0  0
 
-    // Get LGR level
     const int lgr1_level = grid.getLgrNameToLevel().at("LGR1");
     // Total active refined cells on the level grid
     const int numLgrCells = grid.levelGridView(lgr1_level).size(0); // here, 0
@@ -619,7 +459,4 @@ SCHEDULE
 
     // Accessing inactive (non-existing) cell
     BOOST_CHECK_THROW(lgrCartesianIdxToCellIdx.at(0), std::out_of_range);
-
-    // All inactive cells, therefore lgrCOORD(...) throws an exception
-    BOOST_CHECK_THROW(Opm::lgrCOORDandZCORN(grid, lgr1_level, lgrCartesianIdxToCellIdx, lgr1IJK), std::logic_error);
 }
